requirements.txt
pandas
scikit-learn
torch
transformers
peft
matplotlib
seaborn
pytest
pyyaml 
tabulate
wandb
openai
tenacity